NeMo/Code/lib/asr-project/requirements/requirements.txt
pandas
openpyxl
torch
librosa
pytorch_lightning
matplotlib
Cython
nemo_toolkit[all]
